feat: init API branch and readme refresh
