--- api/main.py
@@ -1,85 +1,0 @@
-import os
-
-import dotenv
-import openai
-import requests
-import uvicorn
-import json
-from bs4 import BeautifulSoup
-from fastapi import FastAPI, Response, status
-from fastapi.middleware.cors import CORSMiddleware
-from helpers import database
-from uvicorn.config import LOGGING_CONFIG
-app = FastAPI()
-
-app.add_middleware(
-    CORSMiddleware,
-    allow_origins=["*"],
-    allow_credentials=True,
-    allow_methods=["*"],
-    allow_headers=["*"],
-)
-
-@app.get("/")
-async def root():
-    return {"message": "Alive"}
-
-@app.get("/positions")
-async def get_positions():
-    positions = database.select(query="SELECT * FROM positions")
-    return positions
-
-@app.get("/news/", status_code=200)
-async def get_news(password:str,response: Response, symbol:str = "BTC" ):
-    """
-    Get the latest news about bitcoin from stocktwits
-    """
-    if password != os.getenv("STEIN_API_KEY"):
-        response.status_code = status.HTTP_403_FORBIDDEN
-        return {"message": "Invalid password"}
-    response = requests.get(f"https://stocktwits.com/symbol/{symbol}.X/news")
-    soup = BeautifulSoup(response.content,features="html.parser")
-    news_elements = soup.find_all("div",class_='NewsItem_textContainer__6FGsX')
-    content = str()
-    for index,news in enumerate(news_elements):
-        title = news.find("span", class_="")
-        date = news.find("span",class_="text-light-grey")
-        content+=f"{index+1}. {title.text} - {date.text.split('•')[1]}\n"
-
-    
-
-    dotenv.load_dotenv()
-    openai.api_key = os.getenv("OPENAI_API_KEY")
-
-
-    response = openai.ChatCompletion.create(
-    model="gpt-3.5-turbo",
-    messages=[
-            {
-            "role": "system", 
-            "content": (
-                "Your are a crypto market analyst. Don't take into consideration everything you already know about the crypto market."
-                f"You will be provided with a list of news titles about the {symbol} crypto market."
-                "Your task is to give the overall sentiment as positive, slightly-positive, neutral, slightly-negative or negative."
-                "Then make a general summary of around 250 characters about the news, it will be used for a blog post."
-                "Your answer is a json object, the time key represents the average news time."
-                )
-            },
-            {
-            "role": "assistant", 
-            "content": '{"sentiment" : ..., "summary" : ...,"time": ...}'
-            },
-            {
-            "role": "user",
-            "content": content
-            }
-        ]
-    )
-    
-    print(content)
-
-    return json.loads(response['choices'][0]['message']['content'])
-
-if __name__ == "__main__":
-    LOGGING_CONFIG["formatters"]["access"]["fmt"] = "%(asctime)s [%(name)s] %(levelprefix)s %(message)s"
-    uvicorn.run("main:app", host="0.0.0.0", port=5050, log_level="info",reload=True)
