--- api/requirements.txt
@@ -1,6 +1,0 @@
-fastapi
-uvicorn
-psycopg2-binary
-python-dotenv
-openai
-beautifulsoup4
