--- app/requirements.txt
@@ -1,3 +1,0 @@
-python-dotenv
-requests
-psycopg2-binary
